+++ provisioning/roles/gulp/templates/webpack.config.js
@@ -14,6 +14,19 @@ module.exports = function(production) {
     output: gulpConfig.dest.webpack,
     module: {
       loaders: [
+{% if gulp_use_purescript %}
+        {
+          loader: 'purs-loader',
+          exclude: /node_modules/,
+          test: /\.purs$/,
+          query: {
+            psc: 'psa',
+            pscArgs: { sourceMaps: true },
+            pscIde: true,
+            src: ['bower_components/purescript-*/src/**/*.purs', 'src/**/*.purs']
+          }
+        },
+{% endif %}
         {
           loader: 'babel',
           exclude: /(node_modules|bower_components)/,